Make input path a parameter for WorkBot, refactor states
The work input path is no longer an attribute of the WorkBot (and
should never have been). It (or a WorkInstance) are parameters for its
methods instead.

Removed unnecessary "failed-something" states; we only need to record
the last successful state.

Added new states required for archiving and annotating result data.