Allow WorkBots to provide their own WorkBroker
This change allows configuration to be simplified by allowing a
WorkBot to supply its own broker for finding new work.

This means that to configure a new type of analysis, it is sufficient
to just declare a class for it in the workbot.ini file. The generic
workbot-add and workbot-run scripts will then support all analyses.

It is also possible to pass a broker to a WorkBot constructor, for
cases that need to be customised in code rather than by configuration
file.